1_RAG_csv_pdf/requirements.txt
python-dotenv
langchain
ipykernel
pypdf
langchain_community
langchain-community
bs4
langchain-text-splitters
wikipedia
pymupdf
langchain-openai
chromadb
faiss-cpu
streamlit
langchain_groq
langchain_core
fastapi
langserve
langchain_community
langchain_huggingface
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
